requirements.txt
numpy==2.1.1
py-trees==2.2.3
cython==3.0.11
setuptools==75.1.0
pandas==2.2.3
pyarrow==17.0.0
fastparquet==2024.5.0
ipykernel==6.29.5
jupyter==1.1.1
matplotlib==3.9.2
pygame-ce==2.5.1
seaborn==0.13.2
